Update context data
This is necessary in order to support nominal types and synthesize
eliminators for them. First we need to change the context in order to
tell the difference between builtin, postulates or user-defined values.
Once this is done, we can associate built-in types with a unique
eliminator. From this, we can synthesize eliminators for user-defined
types and associate an eliminator by using the ones from the original
built-in types.